created the ubportal schema

+++ Modules/UBPortal/database/migrations/2026_05_15_092935_create_user_groups_table.php
@@ -6,11 +6,17 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
+    protected $connection = 'ubportal';
     /**
      * Run the migrations.
      */
     public function up(): void
     {
+
+        if (Schema::hasTable('user_groups')) {
+            return;
+        }
+
         // Pivot table mapping users to groups.
         Schema::create('user_groups', function (Blueprint $table) {
             $table->foreignUuid('user_id')->constrained('users')->onDelete('cascade');
