+++ Modules/UBPortal/database/migrations/2026_05_15_093006_create_group_roles_table.php
@@ -4,12 +4,18 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
+    protected $connection = 'ubportal';
     /**
      * Run the migrations.
      */
+
     public function up(): void
     {
+        if (Schema::hasTable('group_roles')) {
+            return;
+        }
         // Pivot table mapping groups to roles.
         Schema::create('group_roles', function (Blueprint $table) {
             $table->foreignUuid('group_id')->constrained('groups')->onDelete('cascade');
